started a git reposetory for my weather app
i have a HTML file, CSS file and a JS file for my project. I have made a fetch API and put it in a function to make it posibel to serch on a city to display the weather as it is right now. I made a fuction for time and date that will show at all time. the HTML is done the Style is almost done and the moste inportent part of JS is done.